refactor(launcher): move MCP test files into src/launcher/mcp/
